LocalPath no longer lowercases it's paths on win32
This commit updates LocalPath to not call str.lower() on paths that are
passed in to it.  The base Path class calls str.lower() before doing any
comparisons already so it seems like this was a bit redundant.

Also updated Path.__hash__ to call str.lower on the path before hashing
when on windows so that behaviour in dictionaries etc. is consistent.