Update switches.py
The proposed changes really represent a choice in approach and are not critical, but I came across a situation where I wanted to provide a list of values for a switch that was generated at run-time. The ultimate goal is that I don't have to maintain the CLI code if I make changes to the program logic that impacts some of the options. I'm reasonably new to Python, so there might be a cleaner pythonic way to handle the situation without modifying the library (plumbum) code--but it was faster and easier for me to make this minor change.