Use composition instead of inheritance in returned Popen objects
The fundamental problem being solved is that Popen objects (quite reasonably)
operate on their own fields. In this case we gave one Popen object's stdin to
another. This solves the problem by exposing all of the fields of an underlying
Popen object via a new class (PlumbumLocalPopen), but any fields we override
are invisible to the underlying object.