SshMachine: add parameter to get remote shell
Sometimes the default remote shell is not compatible with plumbum and the
correct shell needs to be specified.

To add the ability to specify the correct shell the parameter
`remote_shell' can be passed to SshMachine.  If this is passed then the
ssh command is appended with the value.  The arguments passed to
SshMachine.popen are suitably escaped and passed to the remote shell via
the `-c' argument.