Fix bug #395
Even if the child process has exited, it may have written data that is
still in the pipe buffer. We need to make sure to drain it with one
final round of `select()`ing.

Note that this differs from the shell semantics of `| tee file.out`;
In a shell pipeline, the right-hand-side has no direct reference to
the left-hand-side, and so `tee` reads until it gets an `EOF` on the
pipe, normally caused by all write-side ends being closed. We can
demonstrate this with a pipeline that causes the write end to outlive
the left-hand child:

```
( sleep 500 & disown; echo hi ) | tee out.log
```

After the `hi` is printed, we can observe that the left-hand `bash` is
dead, but `tee` still has not returned:

```
$ pstree -p 7273
bash(7273)───tee(1790)
```

Killing the `sleep` makes the whole pipeline exit.

I chose to preserve the original `plumbum` semantics of "exit on child
exit" instead of the shell semantics, but wanted to call out the
difference.