BaseRemoteMachine.tempdir(): add -t to mktemp to work in R/O $HOMEs
The call to mktemp with a relative template name creates the temp
directory wherever the thing logs in to, $HOME.

If $HOME is read-only, then this fails.

With this change, the temp directory is made in whatever the system
considers the temporary directory to be, acknowledging TMPDIR if set
(per mktemp man page) and defaulting to /tmp.